Rewrite react

+++ back/website/services/note_service.py
@@ -8,6 +8,13 @@ class NoteService:
         self.notes_repo = notes_repo
 
     def get_by_id(self, id: int, collection_id: int) -> Response:
+        if id <= 0 or collection_id <= 0:
+            return Response(
+                success=False,
+                message="Id or Collection id cannot be less than or equal to 0",
+                body=None,
+            )
+
         note = self.notes_repo.get_by_id(id=id, collection_id=collection_id)
 
         if note is None:
@@ -15,6 +22,13 @@ class NoteService:
         return Response(success=True, message="Success", body=note)
 
     def get_all(self, collection_id: int) -> Response:
+        if collection_id <= 0:
+            return Response(
+                success=False,
+                message="Collection id cannot be less than or equal to 0",
+                body=None,
+            )
+
         notes = self.notes_repo.get_all(collection_id=collection_id)
 
         if notes is None:
@@ -24,18 +38,36 @@ class NoteService:
     def create_note(self, data: str, user_id: int, collection_id: int) -> Response:
         try:
             if len(data) < 1:
-                return Response(success=False, message="Unable to create", body=None)
+                return Response(
+                    success=False,
+                    message="Unable to create, data is invalid",
+                    body=None,
+                )
+
+            if user_id <= 0 or collection_id <= 0:
+                raise Exception(
+                    "Unable to create, user id or collection id cannot be less than or equal to 0"
+                )
 
             note = self.notes_repo.create_note(
                 data=data, user_id=user_id, collection_id=collection_id
             )
 
             return Response(success=True, message="Successfully created", body=note)
-        except:
-            return Response(success=False, message="Unable to create", body=None)
+        except Exception as ex:
+            return Response(
+                success=False,
+                message=str(ex),
+                body=None,
+            )
 
     def delete_note(self, note_id: int, collection_id: int) -> Response:
         try:
+            if note_id <= 0 or collection_id <= 0:
+                raise Exception(
+                    "Unable to delete, note id or collection id cannot be less than or equal to 0"
+                )
+
             self.notes_repo.delete_note(note_id=note_id, collection_id=collection_id)
 
             return Response(success=True, message="Successfully deleted", body=None)
